sensor mode separation
